% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_consumption.R
\name{BC_consumption}
\alias{BC_consumption}
\title{Consumption data for Body corporate customers}
\usage{
BC_consumption()
}
\arguments{
\item{PROP}{dataframe of consumption from CWW_DW database (dbo.GENTRACK_WATER_EFFICIENCY) in WVDBQDEVSQL}

\item{BC_report}{dataframe of current Body corporate customers produced by the BC_report on the tank}
}
\description{
This takes a dataframe of current Body corporate customers and joins it with the PROP table produced by
the consumption function
}
\examples{
BC_consumption()

}
\keyword{Body}
\keyword{consumption}
\keyword{corporate,}
