% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks}
\alias{breaks}
\title{Consumption breakout function}
\usage{
breaks(DAILYAVG)
}
\arguments{
\item{DAILYAVG}{a vector of daily average consumption}
}
\description{
This takes a vector of consumption and returns the mean consumption of the most recent breakout.
}
\examples{
breaks()

}
\keyword{breakoutDetection}
