% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks2.R, R/breaks3.R
\name{breaks2}
\alias{breaks2}
\alias{breaks2}
\title{Consumption breakout function}
\usage{
breaks2(DAILYAVG)

breaks2(DAILYAVG)
}
\arguments{
\item{DAILYAVG}{a vector of daily average consumption}

\item{DAILYAVG}{a vector of daily average consumption}
}
\description{
This takes a vector of consumption and returns the mean consumption of the most recent breakout.

This takes a vector of consumption and returns the mean consumption of the most recent breakout.
}
\examples{
breaks()

breaks()

}
\keyword{breakoutDetection}
