% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks3.R
\name{breaks3}
\alias{breaks3}
\title{Consumption breakout function}
\usage{
breaks3(DAILYAVG)
}
\arguments{
\item{DAILYAVG}{a vector of daily average consumption}
}
\description{
This takes a vector of consumption and returns the mean consumption of the most recent breakout.
}
\examples{
breaks3()

}
\keyword{breakoutDetection}
