% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks4.R
\name{breaks4}
\alias{breaks4}
\title{Consumption breakout function}
\usage{
breaks4(DAILYAVG)
}
\arguments{
\item{DAILYAVG}{a vector of daily average consumption}
}
\description{
This takes a vector of consumption and returns the mean consumption of the most recent breakout.
}
\examples{
breaks4()

}
\keyword{breakoutDetection}
