% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks_df2.R
\name{breaks_df2}
\alias{breaks_df2}
\title{Consumption breaks dataframe wrapper}
\usage{
breaks_df2(data, column)
}
\arguments{
\item{data}{dataframe of consumption and an ID}

\item{column}{an ID column to pass to the dplyr}
}
\description{
This takes a dataframe of consumption, groups by the column parameter and returns a dataframe of mean consumption of the most
recent breakout by the ID specified.
}
\examples{
breaks_df()

}
\keyword{breakoutDetection,}
\keyword{breaks,}
\keyword{dplyr}
