% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_aves_che.R
\name{build_aves_che}
\alias{build_aves_che}
\title{## Data of Swiss Breeding Birds
Bird presence-absence data from the Swiss breeding bird atlas at a 1x1 km resolution. 
[Schmid, H., Luder, R., Naef-Daenzer, B., Graf, R. & Zbinden, N. (1998) 
Schweizer Brutvogelatlas. Verbreitung der Brutvögel in der Schweiz und
im Fürstentum Liechtenstein 1993-1996. Swiss Ornithological Institute, Sempach, Switzerland].}
\usage{
build_aves_che()
}
\description{
These data were recorded over a four-year period (1993-1996) in usually 
three visits per year (two above the treeline) using a simplified territory 
mapping approach. We only considered species with at least 50 presences resulting 
in a total number of 56 forest bird species in the study region. 
Environmental predictor variables include climate, topography and vegetation structure 
at the same resolution as the species data. Please refer to main article and 
its Online Supplementary Material for description of the data. 
Geographic coordinates were removed.
}
