% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_biotime_che.R
\name{build_biotime_che}
\alias{build_biotime_che}
\title{## Extract biotime data for Switzerland
Please cite as detailed below when using BioTIME:
Dornelas M, Antão LH, Moyes F, Bates, AE, Magurran, AE, et al. 
BioTIME: A database of biodiversity time series for the Anthropocene. 
Global Ecol Biogeogr. 2018; 27:760 - 786. https://doi.org/10.1111/geb.12729
Alternatively, you can load the query from the ZENODO repository
Maria Dornelas, Laura H. Antão, Amanda E. Bates, Viviana Brambilla, Jonathan M. Chase, Cher F. Y. Chow, 
Ada Fontrodona-Eslava, Anne E. Magurran, Inês S. Martins, Faye Moyes, Alban Sagouis, et al. 
BioTIME 2.0: Expanding and Improving a Database of Biodiversity Time Series. 
Global Ecol Biogeogr. 2025; 34(5):e70003. https://doi.org/10.1111/geb.70003}
\usage{
build_biotime_che(filedir)
}
\description{
## Extract biotime data for Switzerland
Please cite as detailed below when using BioTIME:
Dornelas M, Antão LH, Moyes F, Bates, AE, Magurran, AE, et al. 
BioTIME: A database of biodiversity time series for the Anthropocene. 
Global Ecol Biogeogr. 2018; 27:760 - 786. https://doi.org/10.1111/geb.12729
Alternatively, you can load the query from the ZENODO repository
Maria Dornelas, Laura H. Antão, Amanda E. Bates, Viviana Brambilla, Jonathan M. Chase, Cher F. Y. Chow, 
Ada Fontrodona-Eslava, Anne E. Magurran, Inês S. Martins, Faye Moyes, Alban Sagouis, et al. 
BioTIME 2.0: Expanding and Improving a Database of Biodiversity Time Series. 
Global Ecol Biogeogr. 2025; 34(5):e70003. https://doi.org/10.1111/geb.70003
}
