% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ch2018_bioclim_che.R
\name{build_ch2018_bioclim_che}
\alias{build_ch2018_bioclim_che}
\title{---
title: "Calculate CH2018 bioclim data for Switzerland"
author: "RS-eco"
---
## Set-up combinations
## Add time_period to combinations
## Load data and plot
Create output files
Save to file}
\usage{
build_ch2018_bioclim_che(filedir = "inst/extdata")
}
\description{
---
title: "Calculate CH2018 bioclim data for Switzerland"
author: "RS-eco"
---
## Set-up combinations
## Add time_period to combinations
## Load data and plot
Create output files
Save to file
}
