% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cordex_bioclim_che.R
\name{build_cordex_bioclim_che}
\alias{build_cordex_bioclim_che}
\title{---
title: "Calculate Euro-Cordex bioclim data for Switzerland"
author: "RS-eco"
---
### Region (center of boundaries):
~ 27N - 72N, ~ 22W - 45E
### For rotated polar RCMs (in rotated coordinates):
RotPole (198.0; 39.25)
TLC (331.79; 21.67)
Nx=106 Ny=103
### Periods
Hindcast (ERA Interim): 1989 -2008
Control: 1951 - 2005 (1951- 1980, 1981 - 2010)
Scenario: 2006 - 2100 (2011-2040, 2041-71, 2071-2100)
Models for which no files have been downloaded: "CCCma-CanESM2", "ECMWF-ERAINT", "IPSL-IPSL-CM5A-LR", 
"MIROC-MIROC5", "NCC-NorESM1-M", "NOAA-GFDL-GFDL-ESM2G"
## Identify available file combinations
## Subset month period according to 30-yr periods
## Add time_period to combinations
## Remap data and save to file
## Merge data files of subsequent years
Save to file}
\usage{
build_cordex_bioclim_che(filedir)
}
\description{
---
title: "Calculate Euro-Cordex bioclim data for Switzerland"
author: "RS-eco"
---
### Region (center of boundaries):
~ 27N - 72N, ~ 22W - 45E
### For rotated polar RCMs (in rotated coordinates):
RotPole (198.0; 39.25)
TLC (331.79; 21.67)
Nx=106 Ny=103
### Periods
Hindcast (ERA Interim): 1989 -2008
Control: 1951 - 2005 (1951- 1980, 1981 - 2010)
Scenario: 2006 - 2100 (2011-2040, 2041-71, 2071-2100)
Models for which no files have been downloaded: "CCCma-CanESM2", "ECMWF-ERAINT", "IPSL-IPSL-CM5A-LR", 
"MIROC-MIROC5", "NCC-NorESM1-M", "NOAA-GFDL-GFDL-ESM2G"
## Identify available file combinations
## Subset month period according to 30-yr periods
## Add time_period to combinations
## Remap data and save to file
## Merge data files of subsequent years
Save to file
}
