% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_corine_cha_che.R
\name{build_corine_cha_che}
\alias{build_corine_cha_che}
\title{---
title: "Create Corine land cover data for Switzerland"
author: "RS-eco"
---}
\usage{
build_corine_cha_che(filedir)
}
\description{
---
title: "Create Corine land cover data for Switzerland"
author: "RS-eco"
---
}
