% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_forest_che.R
\name{build_forests_che}
\alias{build_forests_che}
\title{Data of Swiss National Forest Inventory}
\usage{
build_forests_che()
}
\description{
Tree species presence-absence data were obtained from the Swiss National Forest Inventory (NFI) . 
The NFI samples Switzerland on a regular grid (spacing 1.4 km), and in case the sample falls into 
forest it records forest characteristics in a maximal area of 50x50 m. 
We aggregated the NFI presence-absence data to 100x100 m plot size to match 
the minimum grain of available environmental data. 
We only considered species with at least 50 presences resulting in a total 
number of 63 tree and shrub species in the study region. 
Environmental predictor variables include climate, topography 
and vegetation structure at the same resolution as the species data. 
Please refer to main article and its Online Supplementary Material 
for description of the data. Geographic coordinates were removed.
}
