% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{ch2018_bioclim_che}
\alias{ch2018_bioclim_che}
\title{Bioclimatic data of Switzerland from CH2018 data}
\format{
A \code{data.frame} with 198240 observations and 25 variables.
}
\description{
Bioclimatic data for current and future conditions in Switzerland
}
\details{
This dataset contains bioclimatic data for Switzerland under 
2 global circulation models (GCMs; MPI-M-MPI-ESM-LR & IPSL-IPSL-CM5A-MR), 
two regional climate models (RCMs; MPI-CSC-REMO2009 & SMHI-RCA4) and 
3 representative concentration pathways (RCPs; RCP2.6, RCP4.5, RCP8.5) 
for current (1991-2020) and future conditions (2041-2070, 2071-2100).
}
