% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{che}
\alias{che}
\title{administrative boundary of Switzerland}
\format{
A \code{data.frame} with 1 observation and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(che)
}
\description{
GADM outline of Switzerland
}
\details{
Administrative boundary of Switzerland derived from GADM data.
}
