% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumption.R
\name{consumption}
\alias{consumption}
\title{Consumption data}
\usage{
consumption(PROP)
}
\arguments{
\item{PROP}{dataframe of consumption from CWW_DW database (dbo.GENTRACK_WATER_EFFICIENCY) in WVDBQDEVSQL}
}
\description{
This takes a dataframe of consumption, groups by the column parameter and returns a dataframe of mean consumption of the most
recent breakout by the ID specified.
}
\examples{
consumption()

}
\keyword{breakoutDetection,}
\keyword{breaks,}
\keyword{dplyr}
