% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{corine_cha_che}
\alias{corine_cha_che}
\title{Corine land cover change data of Switzerland}
\format{
A \code{data.frame} with 15634 observations and 10 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_cha_che)
}
\description{
Corine land cover change data of Switzerland
}
\details{
Corine land cover change data of Switzerland. 
Corine Land cover change are land cover change maps for Europe mapped at a resolution of 100 x 100 m.
}
