% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{corine_lc_che}
\alias{corine_lc_che}
\title{Corine land cover data of Switzerland}
\format{
A \code{data.frame} with 4124593 observations and 7 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://land.copernicus.eu/pan-european/corine-land-cover}}
}
\usage{
data(corine_lc_che)
}
\description{
Corine land cover data of Switzerland
}
\details{
Corine land cover data of Switzerland, Germany. 
Corine Land cover are land cover maps for Europe mapped at a resolution of 100 x 100 m.
}
