% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disposal_class.R
\name{disposal_class}
\alias{disposal_class}
\title{Disposal class table}
\usage{
disposal_class(BC_report)
}
\arguments{
\item{BC_report}{the BC_report returned by getBCdata.R}
}
\description{
This takes the BC_report dataframe and adds in a column for the disposal factor
}
\examples{
disposal_class()

}
