% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalSDC.R
\name{finalSDC}
\alias{finalSDC}
\title{Final SDC}
\usage{
finalSDC(customer)
}
\arguments{
\item{customer}{a row from the BC_report dataframe returned from getBCreport.R}
}
\description{
This takes a customers' disposal class information and summarises it into a "FinalSDC"
}
\examples{
FinalSDC()

}
