% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBCdata.R
\name{getBCData}
\alias{getBCData}
\title{get Body Corp Tank report}
\usage{
getBCData()
}
\description{
This queries Gentrack and gets a report of the active customers with a Body Corp SDC
}
\examples{
getBCdata()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{SRSS}
\keyword{Tank}
\keyword{report,}
