% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBCreport.R
\name{getBCreport}
\alias{getBCreport}
\title{get Body Corp Tank report}
\usage{
getBCreport()
}
\description{
This queries Gentrack and gets a report of the active customers with a Body Corp SDC
}
\examples{
getBCreport()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{SRSS}
\keyword{Tank}
\keyword{report,}
