% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{glarus}
\alias{glarus}
\title{administrative boundary of Canton Glarus, Switzerland}
\format{
A \code{data.frame} with 1 observation and 11 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://www.gadm.org/data.html}}
}
\usage{
data(che)
}
\description{
GADM outline of Canton Glarus, Switzerland
}
\details{
Administrative boundary of Canton Glarus, Switzerland derived from GADM data.
}
