% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{inat_che_1965_2021}
\alias{inat_che_1965_2021}
\title{iNaturalist data of Switzerland}
\format{
A \code{data.frame} object with xxx observations and 9 variables.
}
\source{
This data has been obtained through the rinat package from: 
\itemize{\item \url{https://www.inaturalist.org/}}
}
\usage{
data(inat_che_1965_2021)
}
\description{
iNatrualist data of Switzerland for 1962 - 2022
}
\details{
iNatrualist occurrence data of Switzerland for 1962 - 2022.
}
