% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_one.R
\name{phase_one}
\alias{phase_one}
\title{SDC project Phase one}
\usage{
phase_one()
}
\description{
This reads in a static summary of the results from phase one of the SDC project
}
\examples{
phase_one()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{one}
\keyword{phase}
