% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_two.R
\name{phase_two}
\alias{phase_two}
\title{SDC project Phase two}
\usage{
phase_two()
}
\description{
This reads in a static summary of the results from phase two of the SDC project
}
\examples{
phase_two()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{phase}
\keyword{two}
