% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{potential}
\alias{potential}
\title{potential BC customers to be returned to 0.9}
\usage{
potential()
}
\description{
This reads in a static summary of the results from phase one of the SDC project
}
\examples{
potential()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{one}
\keyword{phase}
