% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential2.R, R/potential3.R
\name{potential2}
\alias{potential2}
\alias{potential2}
\title{potential BC customers to be returned to 0.9}
\usage{
potential2()

potential2()
}
\description{
This reads in a static summary of the results from phase one of the SDC project

This reads in a static summary of the results from phase one of the SDC project
}
\examples{
potential2()

potential2()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{one}
\keyword{phase}
