% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential3.R
\name{potential3}
\alias{potential3}
\title{potential BC customers to be returned to 0.9}
\usage{
potential3()
}
\description{
This reads in a static summary of the results from phase one of the SDC project
}
\examples{
potential3()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{one}
\keyword{phase}
