% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential4.R
\name{potential4}
\alias{potential4}
\title{potential BC customers to be returned to 0.9}
\usage{
potential4()
}
\description{
This reads in a static summary of the results from phase one of the SDC project
}
\examples{
potential4()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{one}
\keyword{phase}
