% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining.R
\name{remaining}
\alias{remaining}
\title{SDC project Phase one}
\usage{
remaining(BC_report)
}
\description{
This reads in a static summary of the results from phase one of the SDC project
}
\examples{
remaining()

}
\keyword{Body}
\keyword{Corporate,}
\keyword{one}
\keyword{phase}
