% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{_PACKAGE}
\name{sdc package}
\alias{sdc package}
\alias{sdcpackage}
\title{R Package to access Swiss climate, elevation, land-use and biodiversity data}
\description{
R Package to access Swiss climate, elevation, land-use and biodiversity data
}
\author{
RS-eco
}
\keyword{package}
