% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{srtm3_che_3arc}
\alias{srtm3_che_3arc}
\title{srtm data of Switzerland}
\format{
A \code{data.frame} with 6999059 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://dwtkns.com/srtm30m/}}
}
\usage{
data(srtm3_che_3arc)
}
\description{
SRTM elevation data of Switzerland
}
\details{
Elevation data of Switzerland derived from NASA SRTM 1 Arc Sec data and aggregated to 3 Arc Sec resolution.
}
