% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{srtm3_glarus_1arc}
\alias{srtm3_glarus_1arc}
\title{srtm data of Glarus, Switzerland}
\format{
A \code{data.frame} with 1060719 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{https://dwtkns.com/srtm30m/}}
}
\usage{
data(srtm3_glarus_1arc)
}
\description{
NASA SRTM v3.0 elevation data of Glarus, Switzerland
}
\details{
Elevation data of Glarus, Switzerland derived from NASA SRTM 1 Arc Sec data.
}
