% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{data}
\name{srtm_csi_che_3arc}
\alias{srtm_csi_che_3arc}
\title{srtm data of Switzerland}
\format{
A \code{data.frame} with 6999059 observations and 3 variables.
}
\source{
This data has been obtained from: \itemize{\item \url{http://srtm.csi.cgiar.org/srtmdata/}}
}
\usage{
data(srtm_csi_che_3arc)
}
\description{
NASA SRTM v3.0 elevation data of Switzerland
}
\details{
Elevation data of Switzerland derived from NASA SRTM 3 Arc Sec v3.0 data.
}
