% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardiseColumns.R
\name{standardiseColumns}
\alias{standardiseColumns}
\title{Standardise Column Names in a Dataframe}
\usage{
standardiseColumns(df, verbose = FALSE)
}
\arguments{
\item{df}{A dataframe whose column names need to be standardized.}

\item{verbose}{A logical indicating whether to print progress messages.}
}
\value{
A dataframe with standardized column names.
}
\description{
This internal function standardizes the column names of a given dataframe.
It uses regular expressions and to match column names against a list of predefined standard names. 
The approach is case-insensitive and allows for flexible matching of column names.
}
\keyword{internal}
