% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{trend}
\alias{trend}
\title{trend the count of Body corporate customers}
\usage{
trend()
}
\description{
This reads in all reports for BC customers and creates a summary dataframe by date
}
\examples{
trend()

}
\keyword{Body}
\keyword{corporate}
