% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writereport.R
\name{writereport}
\alias{writereport}
\title{write report to file}
\usage{
writereport(report, file = getwd(), name)
}
\arguments{
\item{file}{a character string of the file location to write the report to, Defaults to working directory}

\item{name}{a character string of the report name}
}
\description{
This writes a report to file with today's date in the name
}
\examples{
writereport()

}
\keyword{report}
\keyword{write.csv,}
