% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{package}
\name{sdc package}
\alias{sdc}
\alias{sdc package}
\alias{sdc-package}
\title{R Package to access Swiss climate, elevation, land-use and biodiversity data}
\description{
R Package to access Swiss climate, elevation, land-use and biodiversity data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RS-eco/sdc}
  \item Report bugs at \url{https://github.com/RS-eco/sdc/issues}
}

}
\author{
RS-eco
}
\keyword{package}
