% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc-package.R
\docType{package}
\name{sdc package}
\alias{sdc package}
\alias{sdcpackage}
\title{R Package to access Swiss climate, elevation and land-use data}
\description{
R Package to access Swiss climate, elevation and land-use data
}
\author{
RS-eco
}
\keyword{package}
